moved things to common components and refactored
