+++ src/main/java/view/common_components/LabelTextPanel.java
@@ -1,12 +1,12 @@
-package view;
+package view.common_components;
 
 import javax.swing.*;
 
 /**
  * A panel containing a label and a text field.
  */
-class LabelTextPanel extends JPanel {
-    LabelTextPanel(JLabel label, JTextField textField) {
+public class LabelTextPanel extends JPanel {
+    public LabelTextPanel(JLabel label, JTextField textField) {
         this.add(label);
         this.add(textField);
     }
